pyproject.toml
[project]
name = "automata"
version = "0.1.0"
description = "Discord bot for the MUN Computer Science Society"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "beautifulsoup4>=4.12.3",
    "discord-py>=2.4.0",
    "httpx>=0.27.2",
    "jigsaw>=3.2.1",
    "mechanicalsoup>=1.3.0",
    "motor>=3.6.0",
    "nltk>=3.9.1",
    "prometheus-async>=22.2.0",
    "pydantic>=2.9.2",
    "python-dotenv>=1.0.1",
    "pytz>=2024.2",
    "requests>=2.32.3",
    "sentry-sdk>=1.29.2",
]

[dependency-groups]
dev = [
    "ruff>=0.7.3",
]
